Send message to Kafka when a task is updated
