+++ app/controllers/api/v0/service_plans_controller.rb
@@ -28,21 +28,6 @@ module Api
           :provider_control_parameters => {}
         ).to_h
       end
-
-      def messaging_client
-        require "manageiq-messaging"
-
-        @messaging_client ||= ManageIQ::Messaging::Client.open(messaging_opts)
-      end
-
-      def messaging_opts
-        @messaging_opts ||= {
-          :protocol => :Kafka,
-          :host     => ENV["QUEUE_HOST"] || "localhost",
-          :port     => ENV["QUEUE_PORT"] || "9092",
-          :encoding => "json"
-        }
-      end
     end
   end
 end
