+++ app/controllers/api/v0x1.rb
@@ -78,11 +78,6 @@ module Api
       include Api::V0x1::Mixins::IndexMixin
     end
 
-    class TasksController < Api::V0x0::TasksController
-      include Api::V0x1::Mixins::IndexMixin
-      include Api::V0::Mixins::UpdateMixin
-    end
-
     class VmsController < Api::V0x0::VmsController
       include Api::V0x1::Mixins::IndexMixin
     end
